add theme styles, move interface, update test, snapshot
