+++ kafka-ui-react-app/src/components/Nav/ClusterMenuItem/ClusterMenuItem.tsx
@@ -1,10 +1,14 @@
-import StyledMenuItem, {
-  MenuItemProps,
-} from 'components/Nav/ClusterMenuItem/ClusterMenuItem.styled';
+import StyledMenuItem from 'components/Nav/ClusterMenuItem/ClusterMenuItem.styled';
 import React from 'react';
 import { NavLink } from 'react-router-dom';
 
-export type { MenuItemProps };
+export interface MenuItemProps {
+  liType: 'primary';
+  to?: string;
+  activeClassName?: string;
+  title?: string;
+  isInverted?: boolean;
+}
 
 const ClusterMenuItem: React.FC<MenuItemProps> = (props) => {
   const { to, activeClassName, title, children, liType, ...rest } = props;
